moved root emitter out
